Add improved icons

+++ Bakefile.sh
@@ -5,35 +5,38 @@ task.init() {
 }
 
 # shellcheck disable=SC2034
-task.svg() {
+task.generate-assets() {
 	local decoration_color='#ffdeeb'
 	local ecosystem_color='#e5dbff'
 	local platform_color='#ffe8cc'
 
-	magick convert ./assets/package.png -resize 512x512 ./assets/package-512px.png
-	inkscape -w 512 -h 512 -o ./assets/package-512px.png ./assets/package.svg
-	for dir in ./pack-{ecosystem,platform}-!(all|other)/; do
+	for dir in ./pack-*/; do
 		dir=${dir%/}
+		bake.info "Generating for \"$dir\""
 		local icon_name=${dir##*-}
-		local type=${dir#./pack-}; type=${type%-*}
+		local type=${dir#./pack-}; type=${type%%-*}
 		local color_var="${type}_color"
 		local -n color="$color_var"
-		mkdir -p "$dir/assets"
-		magick convert ./assets/package-512px.png -background "$color" -flatten -alpha off ./temp.png
-		if [ ! -f "./assets/$icon_name.png" ]; then
-			if [ -f "./assets/$icon_name.svg" ]; then
-				inkscape -w 512 -o "./assets/$icon_name.png" "./assets/$icon_name.svg"
-			elif [ -f "./assets/$icon_name.jpg" ]; then
-				inkscape -w 512 -o "./assets/$icon_name.png" "./assets/$icon_name.jpg"
-			fi
+
+		if [[ $dir == ./pack-{ecosystem,platform}-!(all|other) ]]; then
+			mkdir -p "$dir/assets"
+			magick convert ./assets/package-512px.png -background "$color" -flatten -alpha off ./temp.png
+			local resize=x225
+			case $icon_name in
+				php) resize=x175 ;;
+				go) resize=x175 ;;
+				java) resize=x325 ;;
+				cpp) resize=x250 ;;
+				latex) resize=x175 ;;
+				markdown) resize=x200 ;;
+				rust) resize=x250 ;;
+				web) resize=x200 ;;
+			esac
+			magick ./temp.png \( "./assets/$icon_name.png" -resize "$resize" -geometry +8+8 \) -gravity southeast -composite "$dir/assets/icon.png"
+			rm ./temp.png
+		else
+			magick convert ./assets/package-512px.png -background "$color" -flatten -alpha off "$dir/assets/icon.png"
 		fi
-		resize=x200
-		[ "$icon_name" = 'php' ] && resize=x150
-		[ "$icon_name" = 'go' ] && resize=x125
-		[ "$icon_name" = 'java' ] && resize=x300
-		[ "$icon_name" = 'cpp' ] && resize=x225
-		magick ./temp.png \( "./assets/$icon_name.png" -resize "$resize" -geometry +8+8 \) -gravity southeast -composite "$dir/assets/icon.png"
-		rm ./temp.png
 	done
 }
 
@@ -43,8 +46,4 @@ task.publish() {
 
 	vsce publish
 	ovsx publish --pat "$(<'../.ovsx-token')"
-}
-
-task.format:fix() {
-	./node_modules/.bin/prettier --write .
 }
